MM-8589: relocate analytics format.jsx to a more common location
