+++ components/analytics/team_analytics/team_analytics.jsx
@@ -13,9 +13,10 @@ import {StatTypes} from 'utils/constants.jsx';
 import Banner from 'components/admin_console/banner.jsx';
 import LineChart from 'components/analytics/line_chart.jsx';
 import StatisticCount from 'components/analytics/statistic_count.jsx';
-import {formatPostsPerDayData, formatUsersWithPostsPerDayData} from 'components/analytics/system_analytics/format.jsx';
 import TableChart from 'components/analytics/table_chart.jsx';
 import LoadingScreen from 'components/loading_screen.jsx';
+
+import {formatPostsPerDayData, formatUsersWithPostsPerDayData} from '../format.jsx';
 
 const LAST_ANALYTICS_TEAM = 'last_analytics_team';
